Moved declaration of version 2 grammar to AFTER Constant Grammar__Version = 2;
The Inform 6.43 code (in progress) imposed restrictions on how
Grammar__Version may be set.  It no longer allows this to be done
anywhere.  It MUST be done before any sort of action is done, otherwise
the compiler will complain.

As a practical matter, no function should appear before the
parser.h file is included.  That's what these changes do.

The commit in the Inform 6 codebase is this:
6a4de5b391379737831ac8ce2081a811c861b9a3